Extract defineInteractor from index
